refactor: 스케줄러의 활성화 토큰 로직을 레디스를 사용하는 로직으로 변경

+++ src/main/java/hhplus/concert/application/userqueue/UserQueueFacade.java
@@ -13,8 +13,7 @@ public class UserQueueFacade {
         return userQueueService.enterUserQueue(concertScheduleId, userId);
     }
     
-    public Integer selectWaitingNumber(String token, Long concertScheduleId) {
+    public Long selectWaitingNumber(String token, Long concertScheduleId) {
         return userQueueService.selectWaitingNumber(token, concertScheduleId);
-
     }
 }
