+++ src/main/java/hhplus/concert/interfaces/api/v1/concert/response/SelectQueueResponse.java
@@ -1,6 +1,6 @@
 package hhplus.concert.interfaces.api.v1.concert.response;
 
 public record SelectQueueResponse(
-        Integer waitingNumber
+    Long waitingNumber
 ) {
 }
